+++ R/tidytab-package.R
@@ -2,5 +2,8 @@
 "_PACKAGE"
 
 ## usethis namespace: start
+#' @importFrom here here
+#' @importFrom readr write_rds
+#' @importFrom utils capture.output
 ## usethis namespace: end
 NULL